Fix weird compiler errors in tests
