--- collections/tests/tests.rs
@@ -1,5 +1,0 @@
-mod segment_tree;
-mod unique_queue;
-
-#[cfg(feature = "bitset")]
-mod bitset;
