+++ primitives/transaction/tests/mod.rs
@@ -15,7 +15,7 @@ const INVALID_EXTENDED_TRANSACTION: &str = "0100004a88aaad038f9b8248865c4b9249ef
 fn it_can_deserialize_extended_transaction() {
     let v: Vec<u8> = hex::decode(EXTENDED_TRANSACTION).unwrap();
     let t: Transaction = Deserialize::deserialize(&mut &v[..]).unwrap();
-    assert_eq!(t.data, Vec::new());
+    assert_eq!(t.data, Vec::<u8>::new());
     assert_eq!(
         t.sender,
         Address::from(&hex::decode("4a88aaad038f9b8248865c4b9249efc554960e16").unwrap()[..])
@@ -55,7 +55,7 @@ fn it_can_serialize_extended_transaction() {
 fn it_can_deserialize_basic_transaction() {
     let v: Vec<u8> = hex::decode(BASIC_TRANSACTION).unwrap();
     let t: Transaction = Deserialize::deserialize(&mut &v[..]).unwrap();
-    assert_eq!(t.data, Vec::new());
+    assert_eq!(t.data, Vec::<u8>::new());
     assert_eq!(
         t.sender,
         Address::from(&hex::decode("b02b9d9fcfa1a60dabe65165ded66a26983404dc").unwrap()[..])
